updated example Structure
